Commit after each file.
This makes the transactions smaller, which ensures that processes
needn't wait very long on each other in case of possible integrity
errors, which might reduce the number of times where a retry is
required. Less waiting means improved parallel utilisation of CPU
time. Although this causes more frequent commits, it seems
reasonable: we're probably hitting the disk once for each file
anyway.

Additionally, this is required to avoid deadlock, otherwise the
sorting we did in 9f8ed3956 doesn't cover all inserts of the
transaction.