Generate pinned and minimum requirements
For testing and deployment purposes, it is very helpful to have pinned
requirements (that won't change when a new release comes out).  It is
also useful to be able to test the purported minimum versions that are
supported, to see if that is in fact true.

This commit does not set up that testing infrastructure, but does add
requirements files, including generated requirements.txt and a new
min-required.txt with minimum claimed supported versions.

This is part of the resolution for Bug 3005.