Drop web interface
It was never used.

From now we are going to use a HTML, CSS, JS approach that will query the
public amaGama API that dynamically builds the results page on the client
browser. Such approach will reduce the size of amaGama and possibly reduce
the server load.

This commit includes a tweak on the requirements to keep the documentation
requirements working properly.