Disable data altering API by default
Some deployments (possibly most of them) won't use this part of the API,
which is in fact most of the API, so disabling it by default might help
quickly deploy with minimum changes.

A setting is used to control if it is enabled or not, so it can be enabled
again if necessary.

Having two separate blueprints for the API makes much more easier to disable
part of the API.