Web UI: Do not use languages on URL for web interface
Now instead of having /en/fr/unit or /en/gl/unit we only have a / URL
for all the possible languages combinations.

This will allow to serve the web UI as static content in a future.

The languages in the dropdowns are automatically populated using
the available source and target languages in the database.