Bug workaround to keep amaGama server alive
When querying for strings like '<a "\b">' (being single quotes and double
quotes part of the string itself) we get a

ProgrammingError: syntax error in tsquery

exception. It is pretty uncommon to query for strings like this, but when
it happens we have the amaGama server down, sometimes for days.