Use a prepared statement for lookup query
This avoids having to build and send the qeury each time in amagama,
and the server only has to parse and and analyze once per connection.
This reduces overhead on all queries by a few milliseconds, depending
on circumstances. It won't help much with the worst queries, but will
speed up the average of all queries a little bit.