Add support for schema-based cursor
This will allow all queries on the cursor to run with the requested schema
in the search_path. We remember the schema used on each connection to not
unnecessarily set it again, otherwise there is an additional query per
call to .get_cursor(). For the case with a single source language and
enough connections in the pool, SET SCHEMA should only need to be run once
per connection per process.