Use separate schema per source language
This way we can stop building names (e.g. sources_%(slang)s). This makes
the SQL in the code usable as-is, which simplifies development, since you
can now simply copy and paste SQL. Names of tables and indexes are
shorter. The schema specified with .get_cursor() is handled such that
there should be no performance impact.

The names of tables and indexes in each schema is the same.