Correctly limit long strings
Source strings longer than the MAX_LENGTH setting will never be used,
unless the setting is changed. Long source or target strings can't
necessarily be indexed successfully during import, and due to compression
and the presence of other columns in the index, there is no simple limit
to test against. This tries to take the MAX_LENGTH setting and the limits
of the btree indexes of PostgreSQL into account and import as much as
possible correctly without bloating the database with things that won't
ever be suggested anyway.

This fixes #3184.