CLUSTER tables for deployment
This improves the physical locality of strings in line with the access
patterns of amagama. Source strings with similar lengths are sought,
and target strings of similar lengths are sought in one language at a
time. This even improved benchmarks with everything in disk cache by
5-15% due to less pages being handled. Cases with cold disk caches should
get a very handsome improvement.

This requires the indexes to limit the search to the relevant pages of
the table. Follow-up commits improve the code further in that area.