Create partially overlapping partial indexes. Performance++
This creates multiple full-text indexes for deployment. It limits the
entries for each search term, thereby helping postgres a bit to limit
the number of rows that need to be retrieved and filtered.

The calculations look at MIN_SIMILARITY and construct the indexes so
that they overlap by exactly the required lengths (but no more) so that
all queries of all lenghts are still handled by the indexes.

Combined with the earlier work on CLUSTERing the tables, this provides a
massive speed improvement, reducing median, average and worse case queries
often by as much as 40% of their previous times.

The hard-coded values (28, 93) were determined experimentally for the
typical dataset of amagama in production. In future this should ideally be
tuned from the database itself. However, it is hard to determine the
"correct" bounds for the indexes since it involves a tradeoff of disk
space vs. improved performance, but increasing the size of the indexes can
also impact performance. This should mostly be an improvement anyway,
particularly for worst cases, since the index used for one query at a time
will be smaller, the number of rows to consider during query execution will
be less, and far less pages are retrieved form disk in the case of cold
caches.

Ideally the last index for the longest strings should also index the
length column. This is postponed until our production server is upgraded
to support the btree_gin extension. Together, that should reduce mean and
95% percentiles to below half their previous timings.